requirements.txt
# === UI Framework ===
streamlit==1.35.0

# === LLM + LangChain Orchestration ===
langchain==0.2.0
langchain-community==0.2.0
openai==1.30.1        # Required for Groq-compatible interface
sentence-transformers==2.6.1        # Required for HuggingFace Embeddings
langchain-community[docloaders]

# === Vector Store ===
qdrant-client==1.8.2

# === Document Loading (wide file support) ===
unstructured==0.13.2      # Handles .txt, .md, .html, etc.
pdfminer.six==20231228    # PDF support
python-docx==1.1.0        # .docx support
beautifulsoup4==4.12.3    # HTML cleaning if needed
PyMuPDF==1.24.2           # Alternative PDF parser for speed

# === Token Counting for context handling ===
tiktoken==0.7.0

# === Env Config / Security ===
python-dotenv==1.0.1
